even more robust updating

+++ startup-scripts/afterUpdate.sh
@@ -3,7 +3,11 @@
 # else you need an extra reboot
 echo "Git changed - copying code across"
 sudo chown pi.pi -R /home/pi/g54mrt-useful-code
-sudo cp /home/pi/g54mrt-useful-code/startup-scripts/rc.local /etc/rc.local
-sudo cp /home/pi/g54mrt-useful-code/startup-scripts/showIP.py /home/pi/showIP.py
+if [ -s "/home/pi/g54mrt-useful-code/startup-scripts/checkUpdate.sh" ] 
+then
+    sudo cp /home/pi/g54mrt-useful-code/startup-scripts/checkUpdate.sh /home/pi/checkUpdate.sh
+    sudo cp /home/pi/g54mrt-useful-code/startup-scripts/rc.local /etc/rc.local
+    sudo cp /home/pi/g54mrt-useful-code/startup-scripts/showIP.py /home/pi/showIP.py
+fi
 cd /home/ubi
 sudo -u ubi bash ./getlatest.sh
